third_party/blink/web_tests/platform/mac-mac14-arm64/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/gpu/parallel-compute.https.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] test parallel WebNN API compute operations
  assert_array_equals: expected property 0 to be 2 but got 8 (expected array [2, 4, 6, 8] got [8, 8, 8, 8])
Harness: the test ran to completion.
